Moved JSON code into source to attempt to fix Windows CI
